backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pandas==2.1.3
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.7
nltk==3.8.1
textblob==0.17.1
openpyxl==3.1.2
cryptograpGGhy==41.0.8
cors==1.0.1
fastapi-cors==0.0.6